requirements.txt
# Core Data Processing
pandas>=1.5.0
numpy>=1.21.0

# Machine Learning
scikit-learn>=1.0.0
tensorflow>=2.0.0

# Data Collection
yfinance>=0.1.70

# NLP & Sentiment Analysis
nltk>=3.6.0
textblob>=0.15.3

# Development & Testing
pytest>=7.0.0
jupyter>=1.0.0
ipykernel>=6.0.0

# Visualization & Configuration
matplotlib>=3.9.2
pyyaml>=6.0.0

# Optional but recommended for deep learning
keras>=2.11.0
h5py>=3.8.0  # For model saving/loading